Code commit - Dec 27 2013

+++ app/Controller/HomeController.php
@@ -54,16 +54,19 @@ class HomeController extends AppController {
  
 	public function display() {
 
-            $this->Session->delete("session_user_id");
-            $this->Session->delete('username');
-            $this->Session->delete('userid');
-            $this->Session->delete('locateid');
-            $this->Session->delete('loginerror');
-            $this->Session->delete('dispstatus');
-            $this->Session->delete('newdiffinfo');
-            $this->Session->delete('newchallengeinfo');
-            $this->Session->delete('stepinfo');
-            $this->Session->destroy();
+if(!$this->Session->read("session_user_id") || $this->Session->read('username'))
+{
+		$this->Session->delete("session_user_id");
+		$this->Session->delete('username');
+		$this->Session->delete('userid');
+		$this->Session->delete('locateid');
+		$this->Session->delete('loginerror');
+		$this->Session->delete('dispstatus');
+		$this->Session->delete('newdiffinfo');
+		$this->Session->delete('newchallengeinfo');
+		$this->Session->delete('stepinfo');
+		$this->Session->destroy();
+}
             
 		$path = func_get_args();
